Refactoring Snap 1
Refactoring Snap 1
