+++ .gitignore
@@ -23,3 +23,5 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 .env
+
+package-lock.json
