+++ src/FAQPage.js
@@ -1,0 +1,11 @@
+import React from "react";
+import FAQ from "./Faq";
+import Grid from "@material-ui/core/Grid";
+
+export default function FAQPage() {
+  return (
+    <Grid item xs={12} sm={12} md={10}>
+      <FAQ />
+    </Grid>
+  );
+}
