+++ src/SelectColumnFilter.js
@@ -1,0 +1,37 @@
+import React from "react";
+
+// This is a custom filter UI for selecting
+// a unique option from a list
+export default function SelectColumnFilter({
+  column: { filterValue, setFilter, preFilteredRows, id }
+}) {
+  // Calculate the options for filtering
+  // using the preFilteredRows
+  const options = React.useMemo(() => {
+    const options = new Set();
+    preFilteredRows.forEach(row => {
+      options.add(row.values[id]);
+    });
+    return [...options.values()];
+  }, [id, preFilteredRows]);
+
+  // Render a multi-select box
+  return (
+    <select
+      className="customStateSelect"
+      value={filterValue}
+      onChange={e => {
+        setFilter(e.target.value || undefined);
+      }}
+    >
+      <option style={{ textAlign: "center" }} value="">
+        All
+      </option>
+      {options.map((option, i) => (
+        <option key={i} value={option}>
+          {option}
+        </option>
+      ))}
+    </select>
+  );
+}
